batch_server/requirements.txt
openai==2.31.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
redis[asyncio]==5.0.8
httpx==0.27.2
pyyaml==6.0.2
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
fakeredis[aioredis]==2.26.1
